Create new router and admin module. Add new postgreSQL connection

+++ app/routers/todos.py
@@ -26,13 +26,21 @@ user_dependency = Annotated[dict, Depends(get_current_user)]
 async def get_todos(user : user_dependency,
                     db: db_dependency):
     if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication credentials were not provided")
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
 
     return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/{todo_id}", status_code=status.HTTP_200_OK)
-async def get_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def get_todo(user : user_dependency,
+                   db: db_dependency,
+                   todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = (db.query(Todos)
+                  .filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
+
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
@@ -41,25 +49,31 @@ class TodoRequest(BaseModel):
     title: str = Field(title="Title", min_length=3)
     description: str = Field(title="Description", min_length=3, max_length=100)
     priority: int = Field(title="Progress", gt=0, lt=6)
-    completed: bool
+    complete: bool
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def create_todo(user : user_dependency,
                       db: db_dependency,
                       todo_request: TodoRequest):
-
     if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication credentials were not provided")
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
 
     todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+async def update_todo(user : user_dependency,
+                      db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = (db.query(Todos)
+                  .filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
+
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
     todo_model.title = todo_request.title
@@ -70,9 +84,17 @@ async def update_todo(db: db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency,
+async def delete_todo(user : user_dependency,
+                      db: db_dependency,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = (db.query(Todos)
+                  .filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id'))
+                  .first())
+
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
     db.delete(todo_model)
